feature: move all currently known supported boards with displays into subpackage
