+++ displays/gopher-badge.go
@@ -1,6 +1,6 @@
 //go:build gopher_badge
 
-package initdisplay
+package displays
 
 import (
 	"image/color"
@@ -10,7 +10,7 @@ import (
 	"tinygo.org/x/tinyterm"
 )
 
-func InitDisplay() tinyterm.Displayer {
+func Init() tinyterm.Displayer {
 	machine.SPI0.Configure(machine.SPIConfig{
 		Frequency: 8000000,
 		Mode:      0,
