+++ displays/wioterminal.go
@@ -1,6 +1,6 @@
-//go:build pyportal
+//go:build wioterminal
 
-package initdisplay
+package displays
 
 import (
 	"image/color"
@@ -10,19 +10,24 @@ import (
 	"tinygo.org/x/tinyterm"
 )
 
-func InitDisplay() tinyterm.Displayer {
-	display := ili9341.NewParallel(
-		machine.LCD_DATA0,
-		machine.TFT_WR,
-		machine.TFT_DC,
-		machine.TFT_CS,
-		machine.TFT_RESET,
-		machine.TFT_RD,
-	)
+func Init() tinyterm.Displayer {
+	machine.SPI3.Configure(machine.SPIConfig{
+		SCK:       machine.LCD_SCK_PIN,
+		SDO:       machine.LCD_SDO_PIN,
+		SDI:       machine.LCD_SDI_PIN,
+		Frequency: 40000000,
+	})
 
 	// configure backlight
-	backlight := machine.TFT_BACKLIGHT
+	backlight := machine.LCD_BACKLIGHT
 	backlight.Configure(machine.PinConfig{machine.PinOutput})
+
+	display := ili9341.NewSPI(
+		machine.SPI3,
+		machine.LCD_DC,
+		machine.LCD_SS_PIN,
+		machine.LCD_RESET,
+	)
 
 	// configure display
 	display.Configure(ili9341.Config{})
