+++ examples/colors/main.go
@@ -7,7 +7,7 @@ import (
 
 	"tinygo.org/x/tinyfont/proggy"
 	"tinygo.org/x/tinyterm"
-	"tinygo.org/x/tinyterm/examples/initdisplay"
+	"tinygo.org/x/tinyterm/displays"
 )
 
 var (
@@ -15,7 +15,7 @@ var (
 )
 
 func main() {
-	display := initdisplay.InitDisplay()
+	display := displays.Init()
 	terminal := tinyterm.NewTerminal(display)
 
 	terminal.Configure(&tinyterm.Config{
